+++ src/Headless.Payments.Paymob.CashIn/Models/Payment/CashInCashCollectionPayResponse.cs
@@ -14,7 +14,7 @@ public sealed class CashInCashCollectionPayResponse
     public required string Pending { get; init; }
 
     [JsonPropertyName("amount_cents")]
-    public int AmountCents { get; init; }
+    public long AmountCents { get; init; }
 
     [JsonPropertyName("success")]
     public required string Success { get; init; }
@@ -60,7 +60,7 @@ public sealed class CashInCashCollectionPayResponse
     public string? TerminalId { get; init; }
 
     [JsonPropertyName("merchant_commission")]
-    public int MerchantCommission { get; init; }
+    public long MerchantCommission { get; init; }
 
     [JsonPropertyName("is_void")]
     public required string IsVoid { get; init; }
@@ -72,10 +72,10 @@ public sealed class CashInCashCollectionPayResponse
     public required string ErrorOccured { get; init; }
 
     [JsonPropertyName("refunded_amount_cents")]
-    public int RefundedAmountCents { get; init; }
+    public long RefundedAmountCents { get; init; }
 
     [JsonPropertyName("captured_amount")]
-    public int CapturedAmount { get; init; }
+    public long CapturedAmount { get; init; }
 
     [JsonPropertyName("merchant_staff_tag")]
     public required string MerchantStaffTag { get; init; }
