+++ src/Headless.Payments.Paymob.CashIn/Models/Payment/CashInSavedTokenPayResponse.cs
@@ -20,7 +20,7 @@ public sealed class CashInSavedTokenPayResponse
     public required string Pending { get; init; }
 
     [JsonPropertyName("amount_cents")]
-    public int AmountCents { get; init; }
+    public long AmountCents { get; init; }
 
     [JsonPropertyName("is_auth")]
     public required string IsAuth { get; init; }
@@ -63,7 +63,7 @@ public sealed class CashInSavedTokenPayResponse
     public string? TerminalId { get; init; }
 
     [JsonPropertyName("merchant_commission")]
-    public int MerchantCommission { get; init; }
+    public long MerchantCommission { get; init; }
 
     [JsonPropertyName("is_void")]
     public required string IsVoid { get; init; }
@@ -75,10 +75,10 @@ public sealed class CashInSavedTokenPayResponse
     public required string ErrorOccured { get; init; }
 
     [JsonPropertyName("refunded_amount_cents")]
-    public int RefundedAmountCents { get; init; }
+    public long RefundedAmountCents { get; init; }
 
     [JsonPropertyName("captured_amount")]
-    public int CapturedAmount { get; init; }
+    public long CapturedAmount { get; init; }
 
     [JsonPropertyName("merchant_staff_tag")]
     public string? MerchantStaffTag { get; init; }
